Fuse fzf.py and main.py files together
