+++ main.py
@@ -1,8 +1,10 @@
 from typing import Literal
 import pathlib
 import sqlite3
-import subprocess
 import sys
+import subprocess
+import enum
+
 
 # TODO: print nice errors instead of python exceptions (do this when we're 40)
 
@@ -80,5 +82,53 @@ def create_titles_with_rating_table():
         conn.commit()
 
 
-if __name__ == "__main__":
+class MainMenuOptions(str, enum.Enum):
+    QUICK_SEARCH = "Quick Search"
+    MOVIES = "Movies"
+    SERIES = "Series"
+    EXPLORE = "Explore"
+
+
+def pipe_into_fzf(lines: list[str], prompt: str = "") -> tuple[str, str]:
+    fzf_proc = subprocess.Popen(
+        ["fzf", "--reverse", "--prompt", prompt],
+        stdin=subprocess.PIPE,
+        stdout=subprocess.PIPE,
+        stderr=subprocess.PIPE,
+        text=True,
+    )
+    stdout, stderr = fzf_proc.communicate(input="\n".join(lines))
+
+    return stdout.rstrip(), stderr.rstrip()
+
+
+def main_menu():
+    choice, _ = pipe_into_fzf(list(MainMenuOptions), prompt="Select an option below: ")
+
+    match choice:
+        case MainMenuOptions.QUICK_SEARCH:
+            # Do a sql search for both movies and series
+            log(f"You selected '{choice}'")
+        case MainMenuOptions.MOVIES:
+            # Do a sql search for movies
+            log(f"You selected '{choice}'")
+        case MainMenuOptions.SERIES:
+            # Do a sql search for series
+            log(f"You selected '{choice}'")
+        case MainMenuOptions.EXPLORE:
+            # This would allow you to browse the database with more control, out of scope for now...
+            log(f"You selected '{choice}'")
+        case "":
+            log("Nothing selected, good bye!")
+            sys.exit(0)
+        case _:
+            raise ValueError(f"'{choice}' is not a valid main menu entry!")
+
+
+def main():
     bootstrap_database()
+    main_menu()
+
+
+if __name__ == "__main__":
+    main()
